Add support for syslogIdentifier. closes #2
This is a port from the 1.x branch (636d623d31f3380b84e221eb37ff67118619426f)

The field SYSLOG_IDENTIFIER is usefull for tagging log messages
with a system name / program name.  It is used on a larger scale
than the logger name, so I think it is usefull to set it as a
property on the appender.